reorg convex modules
