+++ convex/queries.ts
@@ -1,0 +1,50 @@
+import { paginationOptsValidator } from 'convex/server'
+import { v } from 'convex/values'
+import { query } from './_generated/server'
+
+export const list = query({
+  args: {
+    channel: v.string(),
+    limit: v.optional(v.number()),
+  },
+  handler: async (ctx, { channel, limit = 20 }) => {
+    const result = await ctx.db
+      .query('log_entries')
+      .withIndex('channel', (q) => q.eq('channel', channel))
+      .order('desc')
+      .take(Math.min(limit, 100))
+    return result.reverse()
+  },
+})
+
+export const paginate = query({
+  args: {
+    channel: v.string(),
+    paginationOpts: paginationOptsValidator,
+  },
+  handler: async (ctx, { channel, paginationOpts }) => {
+    const result = await ctx.db
+      .query('log_entries')
+      .withIndex('channel', (q) => q.eq('channel', channel))
+      .order('desc')
+      .paginate(paginationOpts)
+    return result
+  },
+})
+
+export const search = query({
+  args: {
+    channel: v.string(),
+    value: v.string(),
+  },
+  handler: async (ctx, args) => {
+    const result = await ctx.db
+      .query('log_entries')
+      .withSearchIndex('messages', (q) =>
+        q.search('content', args.value).eq('channel', args.channel).eq('category', 'message'),
+      )
+      .take(50)
+
+    return result
+  },
+})
