+++ convex/types.ts
@@ -1,0 +1,3 @@
+import type { Doc } from './_generated/dataModel'
+
+export type LogEntry = Doc<'log_entries'>
